backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
alembic==1.14.1
pydantic==2.10.6
PyJWT==2.10.1
python-multipart==0.0.20
argon2-cffi==23.1.0
psycopg[binary]==3.2.6
email-validator==2.2.0

boto3==1.37.38

httpx==0.28.1

prometheus-client==0.21.1
